Decomposed & and fixed some troubles
