+++ src/common/packages/multiple-fetch/tests/response.test.js
@@ -1,4 +1,4 @@
-import {Response} from 'multiple-fetch'
+import {Response} from "../models/response";
 
 const RESOLVE = 'test'
